backend/pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
djangorestframework = "^3.14.0"
rest-framework-roles = "^1.0.3"
django-environ = "^0.11.2"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
djoser = "^2.2.2"
drf-spectacular = "^0.27.1"
markdown = "^3.5.2"
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
